clean up code, move modules to GridImages folder
