+++ client/src/AboutUs.tsx
@@ -4,7 +4,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import { Box, Grid } from '@material-ui/core';
 
 import SmallDisplayCard from './SmallDisplayCard';
-import GridImages from './GridImages';
+import { GridImages } from './DisplayGrids';
 
 import type { Theme } from '@material-ui/core/styles';
