+++ client/src/DisplayGrids/index.ts
@@ -1,0 +1,3 @@
+export { BulletGrid } from './BulletGrid';
+export { InstructionGrid } from './InstructionGrid';
+export { GridImages } from './GridImages';
